+++ packages/install-expo-modules/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 💡 Others
 
+- Updated `AppDelegate.swift` change for SDK 53. ([#36445](https://github.com/expo/expo/pull/36445) by [@kudo](https://github.com/kudo))
+
 ## 0.12.2 — 2025-04-28
 
 _This version does not introduce any user-facing changes._